requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Mail
Flask-WTF
python-dotenv
requests
authlib
pytz
sendgrid
gunicorn
